This gives
```
name    old time/op    new time/op    delta
Read-4    28.4ms ± 1%    26.6ms ± 1%   -6.04%  (p=0.000 n=15+12)

name    old alloc/op   new alloc/op   delta
Read-4    6.90MB ± 0%    0.21MB ± 0%  -96.99%  (p=0.000 n=15+13)

name    old allocs/op  new allocs/op  delta
Read-4       414 ± 0%       316 ± 0%  -23.67%  (p=0.000 n=15+15)
```

and with -conc 3:

```
name    old time/op    new time/op    delta
Read-4    15.5ms ± 1%    14.6ms ± 2%   -5.79%  (p=0.000 n=14+14)

name    old alloc/op   new alloc/op   delta
Read-4    10.8MB ± 0%     4.1MB ± 0%  -61.87%  (p=0.000 n=15+15)

name    old allocs/op  new allocs/op  delta
Read-4       481 ± 0%       383 ± 0%  -20.43%  (p=0.000 n=15+14)
```

in [vcfanno](https://github.com/brentp/vcfanno) that does streaming intersection of many files, this gives ~6-8% improvement in run-time.