@kortschak 2017-01-01
I don't like the use of `io.ReadFull` here. It is intended to signify that the full slice should be filled. The dropping of `io.UnexpectedEOF` is a code smell that follows from its use. If you want to do it this way, write a simplified copy of `io.ReadAtLeast` that just fills the slice and returns n. However, I'd like to know what is causing the iterated read operation to be faster that the `io.Copy` call. This would require profiling.