@brentp 2017-01-01
I can change io.ReadFull to loop over `Read()` until EOF if you prefer, though I don't see why that's preferable. Would a comment above that line indicating the intent of `ReadFull` be sufficient?

Re profiling. Here is readFrom before:
```
(pprof) list readFrom
Total: 331.92MB
ROUTINE ======================== github.com/biogo/hts/bgzf.(*block).readFrom in /home/brentp/go/src/github.com/biogo/hts/bgzf/cache.go
         0   323.96MB (flat, cum) 97.60% of Total
         .          .    127:
         .          .    128:func (b *block) readFrom(r io.ReadCloser) error {
         .          .    129:	o := b.owner
         .          .    130:	b.owner = nil
         .          .    131:	buf := bytes.NewBuffer(b.data[:0])
         .   323.96MB    132:	_, err := io.Copy(buf, r)
         .          .    133:	if err != nil {
         .          .    134:		return err
         .          .    135:	}
         .          .    136:	b.buf = bytes.NewReader(buf.Bytes())
         .          .    137:	b.owner = o
```

and after, that change:
```
(pprof) list readFrom
Total: 2.68MB
```
I'm not sure why io.Copy is allocating so much, I guess it could be re-sizing the slice (not the underlying array) along with the buffer used by io.Copy. Though using io.CopyBuffer with a buffer from a sync.Pool did not improve things as much as this.