@brentp 2017-01-01
On a 3GB VCF, here is memory profile before:
```
(pprof) top10
52.19GB of 52.40GB total (99.59%)
Dropped 14 nodes (cum <= 0.26GB)
      flat  flat%   sum%        cum   cum%
   52.14GB 99.49% 99.49%    52.14GB 99.49%  bytes.makeSlice
    0.05GB 0.098% 99.59%    52.40GB   100%  github.com/biogo/hts/bgzf.(*block).readFrom
         0     0% 99.59%    52.35GB 99.90%  bytes.(*Buffer).ReadFrom
         0     0% 99.59%    52.40GB   100%  github.com/biogo/hts/bgzf.(*decompressor).nextBlockAt.func1
         0     0% 99.59%    52.35GB 99.90%  io.Copy
         0     0% 99.59%    52.35GB 99.90%  io.copyBuffer
         0     0% 99.59%    52.40GB   100%  runtime.goexit
```
and after:
```
(pprof) top10
210.52MB of 210.52MB total (  100%)
Showing top 10 nodes out of 19 (cum >= 3MB)
      flat  flat%   sum%        cum   cum%
  193.02MB 91.69% 91.69%   193.02MB 91.69%  compress/flate.(*huffmanDecoder).init
   14.50MB  6.89% 98.57%   207.52MB 98.57%  github.com/biogo/hts/bgzf.(*block).readFrom
       3MB  1.43%   100%        3MB  1.43%  compress/gzip.(*Reader).readHeader
         0     0%   100%   193.02MB 91.69%  compress/flate.(*decompressor).Read
         0     0%   100%   193.02MB 91.69%  compress/flate.(*decompressor).nextBlock
         0     0%   100%   193.02MB 91.69%  compress/flate.(*decompressor).readHuffman
         0     0%   100%   193.02MB 91.69%  compress/gzip.(*Reader).Read
         0     0%   100%        3MB  1.43%  compress/gzip.(*Reader).Reset
         0     0%   100%        3MB  1.43%  github.com/biogo/hts/bgzf.(*Reader).Read
         0     0%   100%        3MB  1.43%  github.com/biogo/hts/bgzf.(*Reader).nextBlock
```