@kortschak 2017-01-01
Can you look at the profiles in 'io.Copy`? Mainly who is calling makeslice so much.