@brentp 2017-01-02
`bytes.Buffer` implements `ReadFrom` so that is used and that causes the extra allocations. See the graphs below. Also, those and the profiles above are for `-alloc_space`, when using `-alloc_objects` I see a 7% decrease in allocations which matches the performance improvement.
Here is the profile of `ReadFrom`
```
(pprof) list ReadFrom
Total: 52.27GB
ROUTINE ======================== bytes.(*Buffer).ReadFrom in /home/brentp/go/go1.8beta1/go/src/bytes/buffer.go
         0    52.21GB (flat, cum) 99.88% of Total
         .          .    168:			// not enough space at end
         .          .    169:			newBuf := b.buf
         .          .    170:			if b.off+free < MinRead {
         .          .    171:				// not enough space using beginning of buffer;
         .          .    172:				// double buffer capacity
         .    52.01GB    173:				newBuf = makeSlice(2*cap(b.buf) + MinRead)
         .          .    174:			}
         .          .    175:			copy(newBuf, b.buf[b.off:])
         .          .    176:			b.buf = newBuf[:len(b.buf)-b.off]
         .          .    177:			b.off = 0
         .          .    178:		}
         .   208.02MB    179:		m, e := r.Read(b.buf[len(b.buf):cap(b.buf)])
         .          .    180:		b.buf = b.buf[0 : len(b.buf)+m]
         .          .    181:		n += int64(m)
         .          .    182:		if e == io.EOF {
         .          .    183:			break
         .          .    184:		}
```

[before.pdf](https://github.com/biogo/hts/files/680324/before.pdf)
[after.pdf](https://github.com/biogo/hts/files/680323/after.pdf)