@brentp 2017-01-02
[This](https://github.com/golang/go/issues/16474) suggest something to try with using anonymous structs, but that seems even uglier than ReadFull.