@brentp 2017-01-02
I added a helper function that does the loop.