@brentp 2017-01-02
Done.

This is somehow faster. I guess because it can't inline calls with io.Reader args and there were a lot of those now replaced by the single function. On my 3GB test file with -conc 1:
```
benchmark           old ns/op        new ns/op        delta
BenchmarkRead-4     124945818850     107597436845     -13.88%

benchmark           old allocs     new allocs     delta
BenchmarkRead-4     5728424        5324950        -7.04%

benchmark           old bytes       new bytes     delta
BenchmarkRead-4     56388514360     232078760     -99.59%
```
and -conc 3

```
benchmark           old ns/op       new ns/op       delta
BenchmarkRead-4     71698951874     62600825624     -12.69%

benchmark           old allocs     new allocs     delta
BenchmarkRead-4     6135310        5728673        -6.63%

benchmark           old bytes       new bytes       delta
BenchmarkRead-4     86095419064     29939133992     -65.23%
```