@brentp 2017-01-04
addressed and squashed